+++ NEWS.md
@@ -6,13 +6,22 @@ editor_options:
     wrap: 72
 ---
 
-# ggpp 0.5.7
+# ggpp 0.5.7/0.6.0
+
+Track changes in 'ggplot2' 3.5.0. Fix a couple of minor incompatibilities. 
+Add some new features matching those new in 'ggplot2'. General improvements
+and bug fixes.
 
 -   This version depends on 'ggplot2' >= 3.5.0.
+-   Add `as_npc()`, `as_npcx()` and `as_npcy()` helper functions that translate
+positions given as character strings in numeric values in [0..1] into NPC
+(Normalised Parent Coordinates) and validate the range of numeric values if
+passed directly as arguments. The returned values belong to class `AsIs` and
+ready to use in aesthetic mappings.
 -   Update `geom_point_s()` adding parameter `move.point` to allow its use to 
 highlight points at the original position with an arrow with its start given by 
-the action of a position function.
--   Change in `geom_point_s()` the end at which the tip of the arrow is drawn
+the displacement by a position function.
+-   Change in `geom_point_s()` the end at which the tip of the arrow is drawn,
 to its most frequently used direction, even if opposite to the default for
 `geom_text_s()` and `geom_label_s()`.
 -   Update `geom_text_s()`, `geom_label_s()`. `geom_text_pairwise()` and